Channels and slots created and working in AppLayer.cc but problem with multichannel configuration in phy.

+++ src/modules/analogueModel/UWBIRIEEE802154APathlossModel.cc
@@ -195,7 +195,8 @@ void UWBIRIEEE802154APathlossModel::filterSignal(AirFrame *frame, const Coord& /
     // generate number of clusters for this channel (channel coherence time > packet air time)
     L = max(1, poisson(cfg.Lmean));
     // Choose block shadowing
-    S = powf(10.,(normal(0, cfg.sigma_s)/10.));
+    //Changed by Antonio: pow() now -> before powf()
+    S = pow(10.,(normal(0, cfg.sigma_s)/10.));
 
     // Loop on each value of the original mapping and generate multipath echoes
     ConstMappingIterator* iter = txPower->createConstIterator();
